Add `%RELATIVE_CWD_DIR%` and `%RELATIVE_CWD_DOC%` placeholders, move working folder calc to utils
